WeeklyReport/requirements.txt
numpy==1.14.5
pandas==0.23.1
tabula-py==1.3.1
beautifulsoup4==4.6.0